backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "bilalinit", email = "ahmedbilalyt20@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "anthropic>=0.75.0",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.124.0",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "langchain>=1.1.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.0",
    "langchain-openai>=1.1.3",
    "openai>=2.9.0",
    "openai-agents>=0.6.2",
    "openai-chatkit>=1.4.0",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.16.1",
    "sqlalchemy>=2.0.45",
    "tiktoken>=0.12.0",
    "uvicorn>=0.38.0",
]

[project.scripts]
backend = "backend:main"

[build-system]
requires = ["uv_build>=0.9.15,<0.10.0"]
build-backend = "uv_build"

[tool.setuptools]
py-modules = ["backend"]
